Modified scripts to allow reference solvent version

+++ tools/NAMD/postprocess_namd.sh
@@ -8,17 +8,53 @@ failwith() {
     exit 1
 }
 
-args=$#
-if (( args <= 3 )); then
-    failwith "Usage: ./postprocess_namd.sh (psf) (dcd) (xst) (namd log)"
+pos=0
+while true; do
+    arg=$1
+    shift
+    if [[ $arg = "" ]]; then
+	break
+    fi
+    case $arg in
+	(--div=*)
+	DIV=${arg##--div=}
+	;;
+	(--*)
+	failwith "unknown argument $arg"
+	;;
+	(*)
+	case $pos in
+	    (0)
+	    PSF=$arg
+	    ;;
+	    (1)
+	    DCD=$arg
+	    ;;
+	    (2)
+	    LOG=$arg
+	    ;;
+	    (3)
+	    DCDVAC=$arg
+	    ;;
+	    (*)
+	    failwith "too many arguments"
+	    ;;
+	esac
+	(( pos = pos + 1 ))
+	;;
+    esac
+done
+
+if [[ $LOG = "" ]]; then
+    echo "Usage: ./postprocess_namd.sh (psf) (dcd) (namd log) [--div=n]" 1>&2
+    echo "       ./postprocess_namd.sh (psf) (dcd) (namd log) (dcd in vacuo) [--div=n]" 1>&2
+    exit 1
 fi
 
-PSF=$1
-DCD=$2
-XST=$3
-LOG=$4
-
+XST=${DCD%.dcd}.xst
 DCDDIR=${DCD%/*}
+
+# echo PSF $PSF DCD $DCD LOG $LOG DCDVAC $DCDVAC
 
 # check parameter consistency
 
@@ -33,11 +69,15 @@ if [[ ! -e $DCD ]]; then
 fi
 
 if [[ ! -e $XST ]]; then
-    failwith "Error: XST file does not exist!"
+    echo "Warning: XST file does not exist, are you running non-periodic simulation?"
 fi
 
 if [[ ! -e $LOG ]]; then 
     failwith "Error: NAMD log file does not exist!"
+fi
+
+if [[ $DCDVAC != "" ]] && [[ ! -e $DCDVAC ]]; then
+    failwith "Error: DCD file in vacuo is specified but does not exist"
 fi
 
 if [[ ! -e MDinfo ]]; then
@@ -45,16 +85,37 @@ if [[ ! -e MDinfo ]]; then
 fi
 
 if [[ ! -e SltInfo ]]; then
-    failwith "SltInfo file does not exist; try generating it by inpfile.f"
+    failwith "SltInfo file does not exist; try generating it by gen_sltinfo"
 fi
 
-grep "^Info:" $LOG | perl <(cat - <<'EOF'
+# checking condition
+
+if [[ $DCDVAC = "" ]]; then
+    # in solution, no insertion
+    INS=0
+else
+    # in refrence solution system
+    INS=1
+fi
+
+# 10-5 as default
+if [[ $DIV = "" ]]; then
+    if [[ $INS = 0 ]]; then
+	DIV=10
+    else
+	DIV=5
+    fi
+fi
+
+grep "^Info:" $LOG | DIV=$DIV INS=$INS perl <(cat - <<'EOF'
 
 $temp = 300; # as default
 $constant = 1;
 $switchlj = 0;
 $is_periodic = 0;
 $coulombtype = 0;
+if($ENV{INS} == 1){ $instype = 3; } else { $instype = 1; }
+$numdiv=$ENV{DIV};
 while(<>){
 # Info: BERENDSEN PRESSURE COUPLING ACTIVE
 # Info: LANGEVIN PISTON PRESSURE CONTROL ACTIVE
@@ -90,7 +151,7 @@ while(<>){
 }
 
 print "&ene_param
-      slttype = 1,
+      slttype = $instype,
       estype = $constant,
       boxshp = $is_periodic,
       inptemp = $temp,
@@ -104,12 +165,12 @@ print "&ene_param
       ms1max = $box1,
       ms2max = $box2,
       ms3max = $box3,
-      engdiv = 1,
+      engdiv = $numdiv,
       block_threshold = 5.0
 /
 &hist
       eclbin=5.0e-2, ecfbin=2.0e-3, ec0bin=2.0e-4, finfac=10.0e0,
-      ecdmin=-80.0e0, ecfmns=-0.20e0, ecdcen=0.0e0, eccore=20.0e0,
+      ecdmin=-20.0e0, ecfmns=-0.20e0, ecdcen=0.0e0, eccore=20.0e0,
       ecdmax=1.0e11, pecore=200
 /
 "
@@ -122,7 +183,15 @@ EOF
 if [[ -e engsln.01 ]] || [[ -e engsln.tt ]]; then
     echo "Warning: previous output remains, are you running the program twice?" 1>&2
 fi
+
 ln -sf $DCD ./HISTORY
-ln -sf $XST ./HISTCELL
+
+if [[ -e $XST ]]; then
+    ln -sf $XST ./HISTCELL
+fi
+
+if [[ $INS = 1 ]]; then
+    ln -sf $DCDVAC ./SltConf
+fi
